Fixed division math.

+++ Util.py
@@ -3,7 +3,7 @@ import time
 import subprocess
 
 def ceil_to_divisible_by(original: int, divisible_by: int) -> int:
-    return ((original - 1) / divisible_by + 1) * divisible_by
+    return (int((original - 1) / divisible_by) + 1) * divisible_by
 
 def maybe_scale_down_to_fit(current_long_side: int, current_short_side: int, max_long_side: int, max_short_side: int, transposed: bool) -> tuple[int, int]:
     long_side: int = 0
